+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # cypheir-color-scheme
 
+## 0.5.2
+
+### Patch Changes
+
+- worked on a palette mix
+
+## 0.5.1
+
+### Patch Changes
+
+- Added palette set functions
+
+## 0.5.0
+
+### Minor Changes
+
+- Monochrome function tweak - changed logic
+
 ## 0.4.1
 
 ### Patch Changes